create TaskRow
